Refactor token to module
