+++ src/lib.rs
@@ -60,6 +60,7 @@ pub const SDK_VERSION: &str = std::env!("CARGO_PKG_VERSION");
 
 /// The IoT Hub client
 pub mod client;
+
 #[cfg(feature = "http-transport")]
 pub(crate) mod http_transport;
 /// Message types for communicating with the IoT Hub
@@ -67,6 +68,8 @@ pub mod message;
 ///
 #[cfg(not(any(feature = "http-transport", feature = "amqp-transport")))]
 pub mod mqtt_transport;
+///
+pub mod token;
 /// Transport types
 pub mod transport;
